+++ .gitignore
@@ -1,1 +1,2 @@
 /Reto0DIN/build/
+/Reto0DIN/dist/
