+++ Reto0DIN/nbproject/private/private.properties
@@ -1,2 +1,2 @@
 compile.on.save=true
-user.properties.file=C:\\Users\\2dam.TARTANGALH\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
+user.properties.file=C:\\Users\\Ordenador\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
