+++ Reto0DIN/nbproject/private/private.xml
@@ -2,8 +2,6 @@
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/C:/Users/2dam.TARTANGALH/Documents/NetBeansProjects/Reto0_DIN/Reto0DIN/src/view/SwingGraphicViewImplementation.java</file>
-        </group>
+        <group/>
     </open-files>
 </project-private>
