+++ Reto0DIN/src/exception/package-info.java
@@ -4,6 +4,6 @@
  * and open the template in the editor.
  */
 /**
- * In this Package we have the controller 
+ * In this Package we have the Exceptions of the Program 
  */
-package controller;
+package exception;
