+++ Reto0DIN/src/model/FichModelImplementation.java
@@ -7,8 +7,11 @@ package model;
 
 import java.util.ResourceBundle;
 
+
 /**
- *
+ * This class implements the model interface to take the getGreeting Method.
+ * Its one of the 2 classes who is called by the Model Factory
+ * It is connected to a File
  * @author Enaitz Izagirre
  */
 public class FichModelImplementation implements Model{
@@ -16,6 +19,10 @@ public class FichModelImplementation implements Model{
     private ResourceBundle configFile;
     private String greeting;
     
+     /**
+     * This method gets the greeting from the File to return it 
+     * @return Returns the String Greeting from the File
+     */
     @Override
     public String getGreeting() {    
        configFile = ResourceBundle.getBundle("properties.fich");
