+++ Reto0DIN/src/model/package-info.java
@@ -3,4 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
+/**
+ * In this Package we have the Model, his Factory and its implementating variants
+ */
 package model;
