+++ Reto0DIN/src/properties/model.properties
@@ -3,5 +3,5 @@
 # and open the template in the editor.
 
 # Model Type Selection
-Model_Type=BBDD
-#Model_Type=File
+ Model_Type=BBDD
+# Model_Type=File
