+++ Reto0DIN/src/properties/package-info.java
@@ -3,4 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
+/**
+ * In this package we have the .properties configuration files to control the running parameters of the application
+ */
 package properties;
