+++ Reto0DIN/src/properties/view.properties
@@ -4,5 +4,5 @@
 
 # View Type Selection
 #View_Type=Text
-View_Type=Swing
-#View_Type=JavaFX
+#View_Type=Swing
+View_Type=JavaFX
