--- Reto0DIN/src/view/JavaFXGraphicViewAlert.java
@@ -1,34 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package view;
-
-import javafx.scene.control.Alert;
-import javafx.scene.control.Alert.AlertType;
-
-/**
- *
- * @author 2dam
- */
-public class JavaFXGraphicViewAlert {
-
-    public static void infoBox(String infoMessage, String titleBar)
-    {
-        /* By specifying a null headerMessage String, we cause the dialog to
-           not have a header */
-        infoBox(infoMessage, titleBar, null);
-    }
-
-    public static void infoBox(String infoMessage, String titleBar, String headerMessage)
-    {
-        Alert alert = new Alert(AlertType.INFORMATION);
-        alert.setTitle(titleBar);
-        alert.setHeaderText(headerMessage);
-        alert.setContentText(infoMessage);
-        alert.showAndWait();
-    }
-
-
-}
