+++ Reto0DIN/src/view/JavaFXGraphicViewImplementation.java
@@ -5,22 +5,35 @@
  */
 package view;
 
+import javafx.application.Application;
+import javafx.scene.Group;
+import javafx.scene.Scene;
+import javafx.scene.control.Label;
+import javafx.scene.text.Font;
 import javafx.stage.Stage;
 
 /**
  *
  * @author 2dam
  */
-public class JavaFXGraphicViewImplementation implements View {
 
-    private String greeting;
-    private Stage primaryStage;
+public class JavaFXGraphicViewImplementation extends Application implements View {
 
-    @Override
+     @Override
     public void showGreeting(String greeting) {
-        this.greeting = greeting;
-        JavaFXGraphicViewAlert.infoBox(greeting,greeting);
-        
+        String[] parameters= new String [1];
+        parameters[0] = greeting;
+        launch(parameters);        
+    }
+    
+    public void start(Stage primaryStage) {
+        Label text = new Label(getParameters().getRaw().get(0));
+        text.setFont(new Font(40));
+        Scene scene = new Scene(new Group(text));
+        primaryStage.setTitle(getParameters().getRaw().get(0));
+        primaryStage.setScene(scene);
+        primaryStage.sizeToScene();
+        primaryStage.show();
     }
     
 }
